refactor: use a custom hook to retrieve certification info
